refactor: run async setup tasks in parallel & move code & renaming
